+++ documentation/homepage/src/components/Customize/CustomizeTool/Preview/PreviewComponents.svelte
@@ -1,0 +1,12 @@
+<script>
+  import { Column, Grid } from '@flightlesslabs/dodo-ui';
+  import Buttons from './Buttons.svelte';
+</script>
+
+<div>
+  <Grid>
+    <Column>
+      <Buttons />
+    </Column>
+  </Grid>
+</div>