+++ documentation/homepage/src/components/ui/LinkButton.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  import { Button, type ButtonAsAnchorProps } from '@flightlesslabs/dodo-ui';
+
+  let props: ButtonAsAnchorProps = $props();
+</script>
+
+<Button {...props} />